requirements.txt
absl-py==2.1.0
accelerate==0.34.2
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
anywidget==0.9.13
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
arxiv==2.1.3
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
auth==0.5.3
av==13.0.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
black==24.8.0
bleach==6.1.0
blinker==1.8.2
blis==0.7.11
build==1.2.2
cachetools==5.4.0
catalogue==2.0.10
certifi==2024.7.4
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
Chroma==0.2.0
chroma-hnswlib==0.7.6
chromadb==0.5.5
click==8.1.7
cloudpathlib==0.18.1
cloudpickle==3.0.0
cmake==3.30.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
comtypes==1.4.5
confection==0.1.5
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
Cython==3.0.10
dataclasses==0.6
dataclasses-json==0.6.7
datasets==2.20.0
debugpy==1.8.1
decorator==4.4.2
deepface==0.0.92
defusedxml==0.7.1
Deprecated==1.2.14
diffusers==0.30.2
dill==0.3.8
discord==2.3.2
discord.py==2.4.0
distlib==0.3.8
distro==1.9.0
Django==5.1
django-tinymce==4.1.0
dnspython==2.6.1
drawdata==0.3.3
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
eventlet==0.37.0
executing==2.0.1
faiss-cpu==1.8.0.post1
falcon==3.1.3
fastapi==0.115.0
fastjsonschema==2.20.0
feedparser==6.0.11
filelock==3.15.4
fire==0.6.0
flasgger==0.9.7.1
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==24.3.25
fonttools==4.53.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.5.0
future==1.0.0
fvcore==0.1.5.post20221221
gast==0.6.0
gdown==5.2.0
gensim==4.3.3
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud-vision==3.7.4
google-generativeai==0.7.2
google-pasta==0.2.0
googleapis-common-protos==1.65.0
googletrans==3.0.0
greenlet==3.1.0
groq==0.11.0
grpcio==1.66.1
grpcio-status==1.62.3
gTTS==2.5.2
gunicorn==22.0.0
h11==0.14.0
h2==3.2.0
h5py==3.11.0
hpack==3.0.0
hstspreload==2024.9.1
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.24.1
humanfriendly==10.0
hydra-core==1.3.2
hyperframe==5.2.0
idna==2.10
image==1.5.33
imageio==2.34.2
imageio-ffmpeg==0.5.1
imantics==0.1.12
imgaug==0.4.0
importlib_metadata==8.4.0
importlib_resources==6.4.5
imutils==0.5.4
iopath==0.1.10
ipykernel==6.29.4
ipython==8.25.0
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.4.31
jaxlib==0.4.31
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
keras==3.4.1
kiwisolver==1.4.5
kubernetes==30.1.0
labelme2coco==0.2.6
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.1
langchain-google-genai==2.0.0
langchain-groq==0.2.0
langchain-huggingface==0.1.0
langchain-openai==0.2.0
langchain-text-splitters==0.3.0
langcodes==3.4.0
langsmith==0.1.121
language_data==1.2.0
lazy_loader==0.4
libclang==18.1.1
longchain==0.4.2
lxml==5.2.2
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.14
mistune==3.0.2
ml-dtypes==0.4.0
mmh3==5.0.0
mongoengine==0.29.0
monotonic==1.6
MouseInfo==0.1.3
moviepy==1.0.3
mpmath==1.3.0
mtcnn==0.1.1
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
mypy-extensions==1.0.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
neuralintents==0.1.0
nltk==3.8.1
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
omegaconf==2.3.0
onnxruntime==1.19.2
openai==1.45.0
opencv-contrib-python==4.10.0.84
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
opt-einsum==3.3.0
optree==0.12.1
orjson==3.10.7
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pillow==10.4.0
pixellib==0.7.1
platformdirs==4.2.2
plotly==5.22.0
portalocker==2.10.1
posthog==3.6.6
preshed==3.0.9
proglog==0.1.10
prometheus_client==0.20.0
prompt_toolkit==3.0.45
proto-plus==1.24.0
protobuf==4.25.5
psutil==5.9.8
psygnal==0.11.1
pure-eval==0.2.2
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pybboxes==0.1.6
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.5.2
pydantic_core==2.20.1
pydeck==0.9.1
pydot==3.0.1
pygame==2.6.0
PyGetWindow==0.0.9
Pygments==2.18.0
pymongo==4.8.0
PyMsgBox==1.0.9
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pyparsing==3.1.2
pypdf==5.0.0
PyPDF2==3.0.1
pyperclip==1.9.0
PyPika==0.48.9
pypiwin32==223
pyproject_hooks==1.1.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pyreadline3==3.5.3
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pyttsx3==2.90
pytweening==1.2.0
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
qwen-vl-utils==0.0.5
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
retina-face==0.0.17
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.3
sahi==0.11.18
scikit-image==0.24.0
scikit-learn==1.5.0
scipy==1.13.1
screen_brightness_control==0.23.0
seaborn==0.13.2
selenium==4.24.0
Send2Trash==1.8.3
sentence-transformers==3.1.0
sgmllib3k==1.0.0
shapely==2.0.5
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.4.7
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.4
SQLAlchemy==2.0.35
sqlparse==0.5.1
srsly==2.4.8
stack-data==0.6.3
starlette==0.38.5
streamlit==1.38.0
sympy==1.13.2
tabulate==0.9.0
tenacity==8.3.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.18.1
terminaltables==3.1.10
tf_keras==2.17.0
thinc==8.2.5
threadpoolctl==3.5.0
tifffile==2024.7.24
tiktoken==0.7.0
tinycss2==1.3.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.4.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
transformers==4.43.1
trio==0.26.2
trio-websocket==0.11.1
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.12.1
tzdata==2024.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.6
virtualenv==20.26.5
wasabi==1.1.3
watchdog==4.0.1
watchfiles==0.24.0
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.0.1
Werkzeug==3.0.3
widgetsnbextension==4.0.11
wikipedia==1.4.0
WMI==1.5.1
wrapt==1.16.0
wsproto==1.2.0
xmljson==0.2.1
xxhash==3.4.1
yacs==0.1.8
yarl==1.9.4
zipp==3.20.2
